refactor: move withBackOff utility to backoff module and update imports
