+++ packages/notion-api-wrangler/src/notion/database/retrieve.ts
@@ -4,7 +4,7 @@ import type {
   GetDatabaseResponse,
 } from '@notionhq/client/build/src/api-endpoints';
 import type { FetchOptions } from '../../types';
-import { withBackOff } from '../../utils/api';
+import { withBackOff } from '../../utils/backoff';
 
 export const retrieveDatabase = async (
   args: GetDatabaseParameters,
